Separate porting exceptions into their own classes and some other fixes.
