+++ app/src/main/java/org/mercycorps/translationcards/ExportException.java
@@ -1,0 +1,26 @@
+package org.mercycorps.translationcards;
+
+/**
+ * Exception for when there is an error exporting a .txc file.
+ *
+ * @author nick.c.worden@gmail.com (Nick Worden)
+ */
+public class ExportException extends Exception {
+
+    public enum ExportProblem {
+        TARGET_FILE_NOT_FOUND,
+        WRITE_ERROR,
+        TOO_MANY_DUPLICATE_FILENAMES
+    }
+
+    private final ExportProblem problem;
+
+    public ExportException(ExportProblem problem, Throwable cause) {
+        super(cause);
+        this.problem = problem;
+    }
+
+    public ExportProblem getProblem() {
+        return problem;
+    }
+}
